Use retry logic when uploading data.
